Functionality to delete a list from the board.
There is an 'x' button in the top right of a list which once pressed opens a confirmation alert so that you can either cancel or delete the list. With this confirmation there should be less accidental deletions of lists.

I have also reworked a part of the Board view so that lists arrange in a more natural way along the board. This comes with a Scroll Pane so that the user can reach lists that do not fit on the window.